QUAL-253 Refactored tests for elastic client factory
